+++ src/api/messaging.js
@@ -2,8 +2,11 @@ import { handleResponse, handleError } from './apiUtils';
 import { session } from '../models/auth';
 
 export default {
-  getRegion() {
-    const query = `${process.env.REACT_APP_MESSAGING_ROOT}/region`;
+  getRegions(organizationId) {
+    let query = `${process.env.REACT_APP_REGION_API_ROOT}/region`;
+    if (organizationId) {
+      query = `${query}?owner_id=${organizationId}`;
+    }
 
     return fetch(query, {
       method: 'GET',
@@ -15,10 +18,13 @@ export default {
       .then(handleResponse)
       .catch(handleError);
   },
-  getAuthors() {
-    const query = `${process.env.REACT_APP_MESSAGING_ROOT}/author`;
+  async getAuthors(organizationId) {
+    let query = `${process.env.REACT_APP_GROWER_QUERY_API_ROOT}/grower_accounts?author=true`;
+    if (organizationId) {
+      query = `${query}&organization_id=${organizationId}`;
+    }
 
-    return fetch(query, {
+    const res = await fetch(query, {
       method: 'GET',
       headers: {
         'content-type': 'application/json',
@@ -27,20 +33,18 @@ export default {
     })
       .then(handleResponse)
       .catch(handleError);
-  },
-  getAuthorAvatar(handle) {
-    const query = `${process.env.REACT_APP_TREETRACKER_API_ROOT}/grower_accounts?wallet=${handle}`;
 
-    return fetch(query, {
-      method: 'GET',
-      headers: {
-        'content-type': 'application/json',
-        Authorization: session.token,
-      },
-    })
-      .then(handleResponse)
-      .catch(handleError);
+    // interim solution until using try/catch exception for all error messages
+    if (res.error) {
+      throw res.error;
+    }
+
+    const authors = await res.grower_accounts.map((author) => {
+      return { ...author, avatar: author.image_url };
+    });
+    return authors;
   },
+
   postRegion(payload) {
     const query = `${process.env.REACT_APP_MESSAGING_ROOT}/region`;
     const { id, name, description, created_at } = payload;