+++ src/components/PaymentsTable/PaymentsTable.js
@@ -50,7 +50,8 @@ const paymentTableMetaData = [
     description: 'Effective Date',
     name: 'calculated_at',
     sortable: true,
-    showInfoIcon: true,
+    showInfoIcon:
+      'The effective data is the date on which captures were consolidated and the earnings record was created',
   },
   {
     description: 'Payment Method',